+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - New #164: Add `SuggestedPostInfo` type.
 - New #164: Add `suggestedPostInfo` field to `Message` type.
 - New #164: Add `SuggestedPostApproved` type.
+- New #164: Add `suggestedPostApproved` field to `Message` type.
 
 ## 0.9.1 July 4, 2025